+++ docs/CHANGELOG.rst
@@ -6,7 +6,7 @@ All notable changes to the Vector SIL Kit project shall be documented in this fi
 The format is based on `Keep a Changelog (http://keepachangelog.com/en/1.0.0/) <http://keepachangelog.com/en/1.0.0/>`_.
 
 
-[4.0.12] - 2022-11-23
+[4.0.12] - 2022-11-24
 ---------------------
 
 Fixed
@@ -24,6 +24,10 @@ Fixed
 - LIN
 
   - Fixed faulty behavior of ``ILinController::SendFrame()`` for ``LinFrameResponseType::MasterResponse``. The method now consistently uses the input frame data.
+
+- Lifecycle Service
+
+    - Alleviate potential loss of the 'stop' signal issued from the lifecycle service of a participant.
 
 
 [4.0.11] - 2022-11-18